+++ mohaeng/src/main/java/com/mohaeng/alarm/domain/model/AlarmMessageGenerator.java
@@ -1,4 +1,4 @@
-package com.mohaeng.alarm.application.service;
+package com.mohaeng.alarm.domain.model;
 
 import com.mohaeng.alarm.domain.model.value.AlarmMessage;
 import com.mohaeng.alarm.domain.model.value.AlarmType;
