+++ mohaeng/src/test/java/com/mohaeng/alarm/domain/model/RequestClubJoinAlarmContentGeneratorTest.java
@@ -1,6 +1,7 @@
-package com.mohaeng.alarm.application.service.generator;
+package com.mohaeng.alarm.domain.model;
 
 import com.mohaeng.alarm.domain.model.value.AlarmMessage;
+import com.mohaeng.alarm.domain.model.generator.RequestClubJoinAlarmContentGenerator;
 import com.mohaeng.applicationform.domain.event.RequestJoinClubEvent;
 import com.mohaeng.applicationform.domain.model.ApplicationForm;
 import com.mohaeng.applicationform.domain.repository.ApplicationFormRepository;
